+++ SpatialGDK/Source/SpatialGDKFunctionalTests/SpatialGDK/UNR-3761/SpatialTestNetReference/SpatialTestNetReference.h
@@ -29,5 +29,5 @@ public:
 
 	TPair<AController*, APawn*> OriginalPawn;
 
-	float PreviousPositionUpdateFrequency;
+	float PreviousMaximumDistanceThreshold;
 };